requirements.txt
flask==3.0.2
llama-cpp-python==0.2.78
PyPDF2==3.0.1
fpdf2==2.7.8
